Merged software into develop

See merge request tum-ei-esi/wot-team/wot-phyng-simulator!1

+++ backend/python/wopsimulator/openfoam/pyfoam_runner.py
@@ -1,5 +1,5 @@
 import traceback
-from os import kill
+import logging
 from signal import SIGINT
 from threading import Thread, Lock
 
@@ -19,14 +19,18 @@ from PyFoam.Execution.BasicRunner import BasicRunner
 from PyFoam.Execution.ParallelExecution import LAMMachine
 
 
+logger = logging.getLogger('openfoam')
+
+
 class RunFailed(Exception):
     pass
 
 
 def error_callback(error: BaseException):
     """OpenFoam commands error callback"""
-    tb = error.__traceback__
-    traceback.print_exception(type(error), error, tb)
+    pass
+    # tb = error.__traceback__
+    # traceback.print_exception(type(error), error, tb)
 
 
 def run_error_catcher(func):
@@ -42,11 +46,12 @@ def run_error_catcher(func):
         except (RunFailed, Exception) as e:
             error = e
         error_callback(error)
+        raise error
 
     return wrapper
 
 
-def _check_runner_errors(command, solver):
+def check_runner_errors(command, solver):
     """Check if errors appeared after running command"""
     if not solver.runOK():
         if solver.fatalError:
@@ -57,7 +62,10 @@ def _check_runner_errors(command, solver):
             error = f'fatal stack dump'
         else:
             error = 'unknown error'
-        raise RunFailed(f'{command} run failed with {error}: {solver.data["errorText"]}')
+        error_message = f'{command} run failed with {error}' \
+                        f'{(":" + solver.data["errorText"]) if "errorText" in solver.data else ""}'
+        logger.error(error_message)
+        raise RunFailed(error_message)
 
 
 class BasicRunnerWrapper(BasicRunner):
@@ -76,8 +84,10 @@ class BasicRunnerWrapper(BasicRunner):
     def start(self):
         """Starts executing command"""
         self.running = True
-        super(BasicRunnerWrapper, self).start()
-        self.running = False
+        try:
+            super(BasicRunnerWrapper, self).start()
+        finally:
+            self.running = False
 
 
 class PyFoamCmd(BasicRunnerWrapper):
@@ -93,7 +103,7 @@ class PyFoamCmd(BasicRunnerWrapper):
     def start(self):
         """Starts executing command"""
         super(PyFoamCmd, self).start()
-        _check_runner_errors(self.logname, self)
+        check_runner_errors(self.logname, self)
 
 
 class PyFoamSolver(Thread):
@@ -107,32 +117,51 @@ class PyFoamSolver(Thread):
         self._lock = lock
         self._parallel = is_parallel
         self._solver_type = solver_type
-        self._solver = BasicRunnerWrapper(argv=argv, silent=silent, logname=solver_type, lam=lam, **kwargs)
+        self.solver = BasicRunnerWrapper(argv=argv, silent=silent, logname=solver_type, lam=lam, **kwargs)
         super(PyFoamSolver, self).__init__(daemon=True)
 
     @run_error_catcher
     def run(self):
         """Solving thread"""
         with self._lock:
-            print('Entering solver thread')
-            self._solver.start()
-            _check_runner_errors(self._solver_type, self._solver)
-            print('Quiting solver thread')
-            self._solver = None
+            logger.debug('Entering solver thread')
+            try:
+                self.solver.start()
+            except Exception:
+                pass
+            check_runner_errors(self._solver_type, self.solver)
+            logger.debug('Quiting solver thread')
+            self.solver = None
 
     def stop(self, signal):
         """Stops solving"""
-        pid = self._solver.run.run.pid
-        process = psutil.Process(pid)
-        if self._parallel:
+        try:
+            if not self.solver:
+                return
+            pid = self.solver.run.run.pid
+            process = psutil.Process(pid)
             process.children()[0].send_signal(signal)
-        process.send_signal(signal)
+            process.send_signal(signal)
+        except psutil.NoSuchProcess:
+            pass
+        acquired = self._lock.acquire(timeout=1)
+        self._lock.release()
+        if not acquired:
+            logger.warning('Could not obtain the lock, killing the solving thread')
+            self.kill()
 
     def kill(self):
         """Kill solving thread"""
-        self._solver.run.run.send_signal(SIGINT)
-        self._solver = None
         try:
-            self._stop()
-        except AssertionError:
+            pid = self.solver.run.run.pid
+            process = psutil.Process(pid)
+            process.send_signal(SIGINT)
+            process.children()[0].send_signal(SIGINT)
+        except psutil.NoSuchProcess:
             pass
+        self.solver = None
+        acquired = self._lock.acquire(timeout=10)
+        self._lock.release()
+        if not acquired:
+            logger.warning('Solver was not killed within 10 ms')
+            raise Exception('Case solving could not be stopped')
